+++ interface/admin.py
@@ -5,9 +5,8 @@ from pathlib import Path
 from django.contrib import admin, messages
 from django.http import FileResponse
 
-import interface.backend.minio_api as storage
 from interface.models import Course, Assignment, Submission
-from interface.backend.submission import deploy_submission
+
 
 admin.site.register(Course)
 
@@ -52,17 +51,7 @@ class SubmissionAdmin(admin.ModelAdmin):
     actions = ['rerun_submissions']
 
     def rerun_submissions(self, request, submissions):
-        for old_submission in submissions:
-            new_submission = Submission.objects.create(
-                archive_size=old_submission.archive_size,
-                user=old_submission.user,
-                assignment=old_submission.assignment,
-            )
-
-            storage.copy(f'{old_submission.id}.zip',
-                         f'{new_submission.id}.zip')
-
-            deploy_submission(new_submission, old_submission.assignment)
-            new_submission.save()
+        for submission in submissions:
+            submission.evaluate()
 
     rerun_submissions.short_description = 'Re-run submissions'
