+++ interface/utils.py
@@ -1,8 +1,37 @@
+import re
 import string
 import secrets
 import base64
+import configparser
+from urllib.parse import urljoin
+
+import requests
 
 vocabulary_64 = string.ascii_letters + string.digits + '.+'
+
+
+def vmck_config(submission):
+    m = re.match(r'https://github.com/(?P<org>[^/]+)/(?P<repo>[^/]+)/?$',
+                 submission.assignment.repo_url)
+    url_base = ('https://raw.githubusercontent.com'
+                '/{0}/{1}/'.format(*list(m.groups())))
+
+    config_data = requests.get(
+                    urljoin(
+                        url_base,
+                        f'{submission.assignment.repo_branch}/config.ini')
+                        )
+
+    config = configparser.ConfigParser()
+    config.read_string(config_data.text)
+
+    config_dict = dict(config['VMCK'])
+
+    for key, value in config_dict.items():
+        if is_number(value):
+            config_dict[key] = int(value)
+
+    return config_dict
 
 
 def is_number(string):
